refactor(split/t1-b3): mechanically split contract_exit_codes tests
